// web/src/api/models/DtoCreateFamilyMemberRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SubTracker API
 * This api provide HTTPRest endpoints for the application SubTracker.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@mistribe.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DtoCreateFamilyMemberRequest
 */
export interface DtoCreateFamilyMemberRequest {
    /**
     * 
     * @type {Date}
     * @memberof DtoCreateFamilyMemberRequest
     */
    createdAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof DtoCreateFamilyMemberRequest
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof DtoCreateFamilyMemberRequest
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof DtoCreateFamilyMemberRequest
     */
    type: DtoCreateFamilyMemberRequestTypeEnum;
}


/**
 * @export
 */
export const DtoCreateFamilyMemberRequestTypeEnum = {
    Owner: 'owner',
    Adult: 'adult',
    Kid: 'kid'
} as const;
export type DtoCreateFamilyMemberRequestTypeEnum = typeof DtoCreateFamilyMemberRequestTypeEnum[keyof typeof DtoCreateFamilyMemberRequestTypeEnum];


/**
 * Check if a given object implements the DtoCreateFamilyMemberRequest interface.
 */
export function instanceOfDtoCreateFamilyMemberRequest(value: object): value is DtoCreateFamilyMemberRequest {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    return true;
}

export function DtoCreateFamilyMemberRequestFromJSON(json: any): DtoCreateFamilyMemberRequest {
    return DtoCreateFamilyMemberRequestFromJSONTyped(json, false);
}

export function DtoCreateFamilyMemberRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): DtoCreateFamilyMemberRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': json['created_at'] == null ? undefined : (new Date(json['created_at'])),
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'],
        'type': json['type'],
    };
}

export function DtoCreateFamilyMemberRequestToJSON(json: any): DtoCreateFamilyMemberRequest {
    return DtoCreateFamilyMemberRequestToJSONTyped(json, false);
}

export function DtoCreateFamilyMemberRequestToJSONTyped(value?: DtoCreateFamilyMemberRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
        'id': value['id'],
        'name': value['name'],
        'type': value['type'],
    };
}