// web/src/api/models/DtoCreateFamilyRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SubTracker API
 * This api provide HTTPRest endpoints for the application SubTracker.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@mistribe.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DtoCreateFamilyRequest
 */
export interface DtoCreateFamilyRequest {
    /**
     * 
     * @type {Date}
     * @memberof DtoCreateFamilyRequest
     */
    createdAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof DtoCreateFamilyRequest
     */
    creatorName: string;
    /**
     * 
     * @type {string}
     * @memberof DtoCreateFamilyRequest
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof DtoCreateFamilyRequest
     */
    name: string;
}

/**
 * Check if a given object implements the DtoCreateFamilyRequest interface.
 */
export function instanceOfDtoCreateFamilyRequest(value: object): value is DtoCreateFamilyRequest {
    if (!('creatorName' in value) || value['creatorName'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function DtoCreateFamilyRequestFromJSON(json: any): DtoCreateFamilyRequest {
    return DtoCreateFamilyRequestFromJSONTyped(json, false);
}

export function DtoCreateFamilyRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): DtoCreateFamilyRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': json['created_at'] == null ? undefined : (new Date(json['created_at'])),
        'creatorName': json['creator_name'],
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'],
    };
}

export function DtoCreateFamilyRequestToJSON(json: any): DtoCreateFamilyRequest {
    return DtoCreateFamilyRequestToJSONTyped(json, false);
}

export function DtoCreateFamilyRequestToJSONTyped(value?: DtoCreateFamilyRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
        'creator_name': value['creatorName'],
        'id': value['id'],
        'name': value['name'],
    };
}