// web/src/api/models/DtoUpdateFamilyMemberRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SubTracker API
 * This api provide HTTPRest endpoints for the application SubTracker.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@mistribe.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DtoUpdateFamilyMemberRequest
 */
export interface DtoUpdateFamilyMemberRequest {
    /**
     * 
     * @type {string}
     * @memberof DtoUpdateFamilyMemberRequest
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof DtoUpdateFamilyMemberRequest
     */
    type: DtoUpdateFamilyMemberRequestTypeEnum;
    /**
     * 
     * @type {Date}
     * @memberof DtoUpdateFamilyMemberRequest
     */
    updatedAt?: Date;
}


/**
 * @export
 */
export const DtoUpdateFamilyMemberRequestTypeEnum = {
    Owner: 'owner',
    Adult: 'adult',
    Kid: 'kid'
} as const;
export type DtoUpdateFamilyMemberRequestTypeEnum = typeof DtoUpdateFamilyMemberRequestTypeEnum[keyof typeof DtoUpdateFamilyMemberRequestTypeEnum];


/**
 * Check if a given object implements the DtoUpdateFamilyMemberRequest interface.
 */
export function instanceOfDtoUpdateFamilyMemberRequest(value: object): value is DtoUpdateFamilyMemberRequest {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    return true;
}

export function DtoUpdateFamilyMemberRequestFromJSON(json: any): DtoUpdateFamilyMemberRequest {
    return DtoUpdateFamilyMemberRequestFromJSONTyped(json, false);
}

export function DtoUpdateFamilyMemberRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): DtoUpdateFamilyMemberRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'type': json['type'],
        'updatedAt': json['updated_at'] == null ? undefined : (new Date(json['updated_at'])),
    };
}

export function DtoUpdateFamilyMemberRequestToJSON(json: any): DtoUpdateFamilyMemberRequest {
    return DtoUpdateFamilyMemberRequestToJSONTyped(json, false);
}

export function DtoUpdateFamilyMemberRequestToJSONTyped(value?: DtoUpdateFamilyMemberRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'type': value['type'],
        'updated_at': value['updatedAt'] == null ? value['updatedAt'] : value['updatedAt'].toISOString(),
    };
}